sql/schemachanger: move corpus validation into corpus package
Previously, the corpus validation tests were added into end
to end testing to be consistent with master, we are going to
move them into the corpus package.

Release note: None
