+++ pkg/sql/schemachanger/corpus/BUILD.bazel
@@ -1,4 +1,4 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "corpus",
@@ -14,5 +14,19 @@ go_library(
         "//pkg/util/syncutil",
         "@com_github_cockroachdb_errors//:errors",
         "@com_github_cockroachdb_errors//oserror",
+    ],
+)
+
+go_test(
+    name = "corpus_test",
+    srcs = ["corpus_test.go"],
+    tags = ["integration"],
+    deps = [
+        ":corpus",
+        "//pkg/jobs/jobspb",
+        "//pkg/sql/schemachanger/scop",
+        "//pkg/sql/schemachanger/scplan",
+        "//pkg/testutils/skip",
+        "@com_github_stretchr_testify//require",
     ],
 )
